Remove static image generation
